Complete Janki's micro-bench app of using different gpu memory and streams together with the UTC. Write make file and compile successfully on lab-desktop. Not run yet.
some changes to the UTC makefiles.

A bug with using nvcc together with boost1.6, check following for fix:
+// A bug in version 7.0 of CUDA prevents use of variadic templates in
some occasions
+// https://svn.boost.org/trac/boost/ticket/11897
+// This is fixed in 7.5. As the following version macro was introduced
in 7.5 an existance
+// check is enough to detect versions < 7.5
+#if !defined(__CUDACC_VER__) || (__CUDACC_VER__ < 70500)
+#   define BOOST_NO_CXX11_VARIADIC_TEMPLATES
+#endif

+++ utc-0.1/include/thread_util/FastCond.h
@@ -66,7 +66,7 @@ public:
 		if (c.m != m)
 		{
 			/* Atomically set mutex inside cv */
-			cmpxchg(&(c.m), NULL, m);
+			cmpxchg(&(c.m), NULL, (unsigned long long)m);
 			if (c.m != m) return EINVAL;
 		}
